Migrate: Frontend to Tauri 2.0 Beta (#1149)
