+++ rust/tauri-app-win/tauri-app/src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tauri-app"
-version = "0.5.7"
+version = "0.6.0"
 description = "A Tauri App"
 authors = ["you"]
 license = ""
@@ -11,13 +11,17 @@ rust-version = "1.74"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 [build-dependencies]
 tauri-build = { version = "1.5.0", features = [] }
 
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.5.0", features = ["api-all", "icon-ico", "system-tray"] }
+tauri = { version = "1.5.0", features = [ "tray-icon", "api-all", "icon-ico", "system-tray"] }
 
 [features]
 # by default Tauri runs in production mode
